+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.6] - 2022-06-10
+
+### Added
+
+- 💃 `Popover`: added support for custom opener component props
+
+### Fixed
+
+- 📒 `Storybook`: removed `Menu`'s `arg.defaultValue` to get rid of console warning
+
+### Changed
+
+- 📝 `Input`, 🐱‍👤 `MaskedInput`: changed how `ref` is assigned
+
 ## [1.2.5] - 2022-06-09
 
 ### Added